Make several tweaks for Friday Demo and Sprint Review
- un-gray register and login
- (quick fix) --> add in a hard-coded assessmentId to load the criteria
page
- correct current user logic in home page
- Login Page - pretty up styling, add some validation and whistles.
- cast token.jwt as any*
- improve is logged in function
- route all gqls through auth path in constants
- add es2017 to TS rules, prevent annoying casting "errors"

* note: in javascript when using the `+` operator with strings, a
non-string object or primitive will be coerced into a string. So,
regardless if token has a .jwt property or not (either it's going to be
undefined and thus coerced into the string "undefined", or its going to
be some value, which will be coerced into a string). So, I don't need an
error, and I don't need to assume that token itself is going to be a
string, ok Typescript?